test: move gen tests out of src
- Moves the three remaining Gen test files from src to test.
- Updates import paths only; test behavior and runtime code are unchanged.
- Verified with Biome, Gen typecheck, and all 126 tests in the moved files.
